nameservice: using babel instead of esm
