+++ workspaces/nameservice/.gitignore
@@ -1,3 +1,2 @@
-# telepath.config
-idbox-firebase.json
+lib
 Identities.json
