+++ workspaces/nameservice/babel.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  extends: '../../babel.config.js'
+}
