+++ workspaces/nameservice/ecosystem.config.js
@@ -11,7 +11,6 @@ module.exports = {
     },
     env_production: {
       NODE_ENV: 'production'
-    },
-    node_args: '-r esm'
+    }
   }]
 }
