+++ workspaces/nameservice/index.js
@@ -1,48 +1,11 @@
 #!/bin/sh
 ":" //# http://sambal.org/?p=1014 ; exec /usr/bin/env node -r esm "$0" "$@"
 
-import { EntryPoint } from './src/entry-point'
-import commander from 'commander'
+/**
+ * This is a convenience file that allows you to run the service in esm mode - so
+ * without babel transpilation in between
+ */
 
-const program = new commander.Command()
-
-const start = cmdObj => {
-  const { servicePath, registrationPath } = cmdObj
-  console.log('servicePath=', servicePath)
-  console.log('registrationPath=', registrationPath)
-  const entryPoint = new EntryPoint({
-    servicePath,
-    registrationPath
-  })
-  entryPoint.start()
-  process.on('SIGINT', () => {
-    console.log(`stopping ${servicePath}...`)
-    entryPoint.stop()
-    console.log('stopped. exiting now...')
-    process.exit(0)
-  })
-  process.stdin.resume()
-}
-
-async function main () {
-  program
-    .version('0.1.0')
-    .usage('command [options]')
-    .on('command:*', () => {
-      console.error('Invalid command: %s\nSee --help for a list of available commands.', program.args.join(' '))
-      process.exit(1)
-    })
-
-  program.command('start')
-    .option('-p, --servicePath <path>', 'service path for the service in the format: service-namespace.service-id', 'identity-box.nameservice')
-    .option('-r, --registrationPath <path>', 'registration path for the service in the format: service-namespace.service-id', 'identity-box.service-registration')
-    .action(start)
-
-  await program.parse(process.argv)
-
-  if (!process.argv.slice(2).length) {
-    program.help()
-  }
-}
+import { main } from './source/main'
 
 main()
