+++ workspaces/nameservice/source/entry-point/EntryPoint.js
