+++ workspaces/nameservice/source/entry-point/index.js
