+++ workspaces/nameservice/source/index.js
@@ -1,0 +1,5 @@
+#!/usr/bin/env node
+
+import { main } from './main'
+
+main()
