+++ workspaces/nameservice/source/services/NamePublisher.js
