+++ workspaces/nameservice/source/services/NameResolver.js
