+++ workspaces/nameservice/source/services/NameService.js
