+++ workspaces/nameservice/tools/build.js
@@ -1,0 +1,5 @@
+const { Builder } = require('../../../tools/build')
+
+const builder = new Builder()
+
+builder.build()
